# docs/api/v1/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5078/"
    }
  ],
  "paths": {
...
    "/api/rules": {
      "post": {
        "tags": [
          "Rules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "get": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/rules/{id}": {
      "get": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdateRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdateRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rules/{id}/activate": {
      "post": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rules/{id}/deactivate": {
      "post": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rules/{id}/promote": {
      "post": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePromotionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RulePromotionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RulePromotionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rules/{id}/dry-run": {
      "post": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/rules/{id}/rollback": {
      "post": {
        "tags": [
          "Rules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleRollbackRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleRollbackRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleRollbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/rules/test-fixture": {
      "post": {
        "tags": [
          "Rules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleFixtureTestRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleFixtureTestRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleFixtureTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/rules/events/dispatch": {
      "post": {
        "tags": [
          "Rules"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTriggerEventRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTriggerEventRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTriggerEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "RuleActionRequest": {
        "required": [
          "type",
          "value"
        ],
        "type": "object",
        "properties": {
          "type": {
            "maxLength": 100,
            "type": "string"
          },
          "value": {
            "maxLength": 200,
            "type": "string"
          }
        }
      },
      "RuleConditionRequest": {
        "required": [
          "field",
          "operator",
          "value"
        ],
        "type": "object",
        "properties": {
          "field": {
            "maxLength": 100,
            "type": "string"
          },
          "operator": {
            "maxLength": 30,
            "type": "string"
          },
          "value": {
            "maxLength": 200,
            "type": "string"
          }
        }
      },
      "RuleCreateRequest": {
        "required": [
          "name",
          "trigger"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 160,
            "type": "string"
          },
          "trigger": {
            "maxLength": 100,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "priority": {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "conflictPolicy": {
            "maxLength": 20,
            "type": "string"
          },
          "executionStartHourUtc": {
            "maximum": 23,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "executionEndHourUtc": {
            "maximum": 23,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cooldownMinutes": {
            "maximum": 10080,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "allowDestructiveActions": {
            "type": "boolean"
          },
          "environment": {
            "maxLength": 12,
            "type": "string"
          },
          "approvalStatus": {
            "maxLength": 20,
            "type": "string"
          },
          "conditions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleConditionRequest"
            }
          },
          "actions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleActionRequest"
            }
          }
        }
      },
      "RuleFixtureLeadRequest": {
        "type": "object",
        "properties": {
          "source": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "score": {
            "type": "integer",
            "format": "int32"
          },
          "hasEmail": {
            "type": "boolean"
          },
          "hasPhone": {
            "type": "boolean"
          },
          "fromStage": {
            "type": "string",
            "nullable": true
          },
          "toStage": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "RuleFixtureTestRequest": {
        "required": [
          "ruleId",
          "trigger",
          "lead"
        ],
        "type": "object",
        "properties": {
          "ruleId": {
            "type": "string",
            "format": "uuid"
          },
          "trigger": {
            "maxLength": 100,
            "type": "string"
          },
          "lead": {
            "$ref": "#/components/schemas/RuleFixtureLeadRequest"
          }
        }
      },
      "RulePromotionRequest": {
        "required": [
          "targetEnvironment",
          "approvedBy"
        ],
        "type": "object",
        "properties": {
          "targetEnvironment": {
            "maxLength": 12,
            "type": "string"
          },
          "approvedBy": {
            "maxLength": 160,
            "type": "string"
          }
        }
      },
      "RuleRollbackRequest": {
        "type": "object",
        "properties": {
          "targetVersion": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
      "RuleTriggerEventRequest": {
        "required": [
          "trigger"
        ],
        "type": "object",
        "properties": {
          "trigger": {
            "maxLength": 100,
            "type": "string"
          },
          "leadId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "opportunityId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "proposalId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fromStage": {
            "maxLength": 80,
            "type": "string",
            "nullable": true
          },
          "toStage": {
            "maxLength": 80,
            "type": "string",
            "nullable": true
          }
        }
      },
      "RuleUpdateRequest": {
        "required": [
          "name",
          "trigger"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 160,
            "type": "string"
          },
          "trigger": {
            "maxLength": 100,
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "priority": {
            "maximum": 1000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "conflictPolicy": {
            "maxLength": 20,
            "type": "string"
          },
          "executionStartHourUtc": {
            "maximum": 23,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "executionEndHourUtc": {
            "maximum": 23,
            "minimum": 0,
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cooldownMinutes": {
            "maximum": 10080,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "allowDestructiveActions": {
            "type": "boolean"
          },
          "environment": {
            "maxLength": 12,
            "type": "string"
          },
          "approvalStatus": {
            "maxLength": 20,
            "type": "string"
          },
          "conditions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleConditionRequest"
            }
          },
          "actions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RuleActionRequest"
            }
          }
        }
      },
...
    }
  },
...
}